Complete the page-run cleanup audit
