+++ swath-cli/src/main/java/io/varve/swath/cli/TuneOptions.java
@@ -8,7 +8,6 @@ package io.varve.swath.cli;
 import io.varve.swath.error.InvalidArgsException;
 import io.varve.swath.error.InvalidConfigException;
 import io.varve.swath.sort.SortConfig;
-import io.varve.swath.sort.SortFinalization;
 import io.varve.swath.sort.StagingRetention;
 import java.io.PrintWriter;
 import java.time.Duration;
@@ -37,9 +36,6 @@ final class TuneOptions {
                             + SortOptions.MAX_MERGE_PARALLELISM,
                     Integer.toString(SortConfig.DEFAULT.mergeParallelism()), "stable", ResumeClass.FREE,
                     "fresh list and resume"),
-            new KeySpec(SortConfig.FINALIZATION_TUNE_KEY, "enum", "pipeline",
-                    SortConfig.DEFAULT.finalization().configValue(), "experimental",
-                    ResumeClass.FREE, "fresh list and resume"),
             new KeySpec(SortConfig.KEEP_STAGING_TUNE_KEY, "boolean", "on|off",
                     SortConfig.DEFAULT.stagingRetention().tuneValue(), "diagnostic", ResumeClass.FREE,
                     "fresh list and resume"),
@@ -102,9 +98,6 @@ final class TuneOptions {
             } else if ("sort.merge-parallelism".equals(key)) {
                 sorting.mergeParallelism = parseMergeParallelism(key, setting.getValue());
                 effectiveValues.put(key, Integer.toString(sorting.mergeParallelism));
-            } else if (SortConfig.FINALIZATION_TUNE_KEY.equals(key)) {
-                sorting.finalization = parseFinalization(key, setting.getValue());
-                effectiveValues.put(key, sorting.finalization.configValue());
             } else if (SortConfig.KEEP_STAGING_TUNE_KEY.equals(key)) {
                 sorting.stagingRetention = StagingRetention.fromEnabled(parseOnOff(key, setting.getValue()));
                 effectiveValues.put(key, sorting.stagingRetention.tuneValue());
@@ -123,8 +116,6 @@ final class TuneOptions {
                                 ? resolvedSummaryInterval
                                 : "sort.merge-parallelism".equals(spec.key())
                                         ? Integer.toString(resolvedSortConfig.mergeParallelism())
-                                        : SortConfig.FINALIZATION_TUNE_KEY.equals(spec.key())
-                                                ? resolvedSortConfig.finalization().configValue()
                                         : SortConfig.KEEP_STAGING_TUNE_KEY.equals(spec.key())
                                                 ? resolvedSortConfig.stagingRetention().tuneValue()
                                         : spec.defaultValue()))
@@ -228,10 +219,6 @@ final class TuneOptions {
                     sorting.mergeParallelism = parseMergeParallelism(key, value);
                     effectiveValues.put(key, Integer.toString(sorting.mergeParallelism));
                 }
-                case SortConfig.FINALIZATION_TUNE_KEY -> {
-                    sorting.finalization = parseFinalization(key, value);
-                    effectiveValues.put(key, sorting.finalization.configValue());
-                }
                 case SortConfig.KEEP_STAGING_TUNE_KEY -> {
                     sorting.stagingRetention = StagingRetention.fromEnabled(parseOnOff(key, value));
                     effectiveValues.put(key, sorting.stagingRetention.tuneValue());
@@ -252,15 +239,6 @@ final class TuneOptions {
             throw valueError(key, value);
         }
         return parallelism;
-    }
-
-    private static SortFinalization parseFinalization(String key, String value)
-            throws InvalidArgsException {
-        try {
-            return SortFinalization.fromConfigValue(key, value);
-        } catch (IllegalArgumentException e) {
-            throw valueError(key, value);
-        }
     }
 
     private static boolean parseOnOff(String key, String value) throws InvalidArgsException {
