+++ swath-cli/src/test/java/io/varve/swath/cli/TuneOptionsTest.java
@@ -8,7 +8,6 @@ package io.varve.swath.cli;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.varve.swath.sort.SortConfig;
-import io.varve.swath.sort.SortFinalization;
 import io.varve.swath.sort.StagingRetention;
 import java.io.PrintWriter;
 import java.io.StringWriter;
@@ -19,7 +18,6 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.parallel.ResourceLock;
 import picocli.CommandLine;
 
 class TuneOptionsTest {
@@ -32,7 +30,6 @@ class TuneOptionsTest {
                 "--tune", "parquet.writers=4",
                 "--tune", "summary.interval=2s",
                 "--tune", "sort.merge-parallelism=16",
-                "--tune", "sort.finalization=pipeline",
                 "--tune", "sort.keep-staging=on",
                 "--tune", "sort.ignore-disk-check=on");
 
@@ -43,7 +40,6 @@ class TuneOptionsTest {
         assertThat(cmd.output.parquetWriters).isEqualTo(4);
         assertThat(cmd.resolveSummaryJsonInterval()).isEqualTo(Duration.ofSeconds(2));
         assertThat(cmd.sorting.resolveConfig().mergeParallelism()).isEqualTo(16);
-        assertThat(cmd.sorting.resolveConfig().finalization()).isEqualTo(SortFinalization.PIPELINE);
         assertThat(cmd.sorting.resolveConfig().stagingRetention().retainsOriginals()).isTrue();
         assertThat(cmd.sorting.forceSort).isTrue();
     }
@@ -73,7 +69,6 @@ class TuneOptionsTest {
                 {"summary.interval=zero", "positive duration"},
                 {"sort.merge-parallelism=0", "integer 1..16"},
                 {"sort.merge-parallelism=17", "integer 1..16"},
-                {"sort.finalization=lanes", "pipeline"},
                 {"sort.keep-staging=yes", "on|off"},
                 {"sort.ignore-disk-check=yes", "on|off"},
         };
@@ -146,7 +141,6 @@ class TuneOptionsTest {
                         "tune effective: engine.readahead=off, seed.mode=shallow, "
                         + "parquet.writers=4, summary.interval=PT30S, "
                         + "sort.merge-parallelism=" + SortConfig.DEFAULT.mergeParallelism() + ", "
-                        + "sort.finalization=pipeline, "
                         + "sort.keep-staging=off, "
                         + "sort.ignore-disk-check=off");
     }
@@ -172,28 +166,6 @@ class TuneOptionsTest {
 
         assertThat(tune.applyForResume(sorting, new PrintWriter(new StringWriter()))).isFalse();
         assertThat(sorting.resolveConfig().mergeParallelism()).isEqualTo(16);
-    }
-
-    @Test
-    @ResourceLock("SYSTEM_PROPERTIES")
-    void finalizationIsResumeSafe() throws Exception {
-        String property = SortConfig.FINALIZATION_PROPERTY;
-        String previous = System.getProperty(property);
-        try {
-            System.setProperty(property, "pipeline");
-            TuneOptions tune = new TuneOptions();
-            tune.entries = List.of(SortConfig.FINALIZATION_TUNE_KEY + "=pipeline");
-            SortOptions sorting = new SortOptions();
-
-            assertThat(tune.applyForResume(sorting, new PrintWriter(new StringWriter()))).isFalse();
-            assertThat(sorting.resolveConfig().finalization()).isEqualTo(SortFinalization.PIPELINE);
-        } finally {
-            if (previous == null) {
-                System.clearProperty(property);
-            } else {
-                System.setProperty(property, previous);
-            }
-        }
     }
 
     @Test
