+++ swath-core/src/main/java/io/varve/swath/sort/CascadePageMerger.java
@@ -17,8 +17,7 @@ import java.util.PriorityQueue;
 /**
  * Page-whole cascade merger for page-run inputs. A page is streamed directly when its maximum is
  * below every successor frontier; only a transitively overlapping component enters the shared
- * {@link PageRowMerger}. This keeps the surviving cascade path off the per-row heap for disjoint
- * pages without restoring the deleted range/seek frontier abstractions.
+ * {@link PageRowMerger}. Disjoint pages therefore stay off the per-row heap.
  */
 final class CascadePageMerger implements SortedCursor, LogicalMergeCompletion {
