+++ swath-core/src/main/java/io/varve/swath/sort/SortedCursor.java
@@ -11,8 +11,8 @@ import java.util.Iterator;
 /**
  * A forward-only stream of {@link ListEntry} in {@link ListEntryComparator} order — returned by a
  * fixture-staging chunk or the {@link SortTransform} merge phase. Backed either by an in-memory
- * sorted list or by a streaming k-way merge over open {@link EntryStream}s,
- * so it is {@link AutoCloseable}: use it in a try-with-resources to release any open readers.
+ * sorted list or by the page-run merge pipeline, so it is {@link AutoCloseable}: use it in a
+ * try-with-resources to release any open readers.
  *
  * <p>Underlying read failures surface as {@link java.io.UncheckedIOException} from {@link #next()}
  * (the {@link Iterator} contract forbids checked exceptions).
