+++ swath-core/src/test/java/io/varve/swath/sort/PageRunReads.java
@@ -5,15 +5,16 @@
  */
 package io.varve.swath.sort;
 
+import io.varve.swath.model.ListEntry;
 import java.io.IOException;
+import java.io.UncheckedIOException;
 import java.nio.file.Path;
+import java.util.ArrayList;
 import java.util.List;
 
 /**
- * Test-only reader for page-run staging segments, in-package so it can reach the package-private
- * {@link PageRunSegmentReader}. Exposed publicly so tests in other packages (e.g. the runtime
- * SORT-RESUME adversarial tests) can read staging keys back without a Parquet reader — the staging format is
- * page-run, not columnar Parquet.
+ * Test-only row oracle for page-run staging segments. It deliberately drives the production
+ * {@link PageRunSegmentIo#nextPage()} loop instead of maintaining a second segment reader.
  */
 public final class PageRunReads {
 
@@ -22,13 +23,31 @@ public final class PageRunReads {
 
     /** The keys of a page-run staging/intermediate segment, in stored (sorted) order. */
     public static List<String> keys(Path segment) throws IOException {
-        return SortTestSupport.drain(open(segment)).stream()
+        return entries(segment).stream()
                 .map(entry -> entry.key().asString())
                 .toList();
     }
 
-    static PageRunSegmentReader open(Path segment) throws IOException {
-        return new PageRunSegmentReader(
-                PageRunSegmentIo.open(segment, SortMetrics.NO_OP), new ListEntryComparator());
+    static List<ListEntry> entries(Path segment) throws IOException {
+        return entries(segment, SortMetrics.NO_OP);
+    }
+
+    static List<ListEntry> entries(Path segment, SortMetrics metrics) throws IOException {
+        List<ListEntry> result = new ArrayList<>();
+        try {
+            try (PageRunSegmentIo io = PageRunSegmentIo.open(segment, metrics)) {
+                PageRunSegmentIo.Page page;
+                while ((page = io.nextPage()) != null) {
+                    PageBlockCursor cursor = page.decode(segment).cursor();
+                    while (cursor.hasNext()) {
+                        result.add(cursor.next());
+                    }
+                    cursor.drainAndValidate();
+                }
+            }
+        } catch (UncheckedIOException failure) {
+            throw failure.getCause();
+        }
+        return result;
     }
 }
