+++ swath-core/src/test/java/io/varve/swath/sort/SortLaneTest.java
@@ -127,13 +127,6 @@ final class SortLaneTest {
     }
 
     private List<String> keys(Path segment) throws IOException {
-        // The lane stages page-run (.pageseg) segments; read them back with the page-run reader.
-        List<String> out = new ArrayList<>();
-        try (PageRunSegmentReader r = PageRunReads.open(segment)) {
-            while (r.hasNext()) {
-                out.add(r.next().key().asString());
-            }
-        }
-        return out;
+        return PageRunReads.keys(segment);
     }
 }
